now the new credentials are in the credentials repository
We had to manually add the bean dependency. This is because a plugin is isolated in its own classloader and Spring context, with the app's Spring context as the parent. Components from the app will automatically get autowired into the plugin. And if an app component is autowired with a single bean, and the plugin defines that bean as primary, a dependency is formed implicitly. But if an app component takes a collection of beans and that is already satisfied, it isn't able to wait to see if more are in the plugin. So you have to manually create a dependency.
We do this by overriding registerBeanDefinitions which the the method that does most of the work in SpringLoaderPlugin. In this case we manually make artifactCredentialsRepository depend on the plugin. That was it guarantees that the plugin beans will be initialized before artifactCredentialsRepository and the new credentials will be available.